chore: refactor, add test
