--- packages/legacy/core/App/types/help.ts
@@ -1,10 +1,0 @@
-import { StackNavigationProp } from '@react-navigation/stack'
-
-import { NotificationStackParams, Screens } from './navigators'
-
-export type HelpAction = (navigation: StackNavigationProp<NotificationStackParams, Screens.ProofRequest>) => void
-
-export interface HelpEntry {
-  credDefIds: string[]
-  action: HelpAction
-}
